General framework for robot failures
Robots are assigned whether or not they will fail
If applicalbe, robots are then assigned what time they will fail
When the robot does fail it it checks which failure case it is supposed to do.

This is a general framework as the robot does not implement the specific failure cases yet.