Changed method of picking which robots to fail
Through the argos file you can now set the following:
failure_rate -- percentage of robots to fail
maxFail -- The latest that a robot is allowed to fail

Also please make sure that num_robots is always the same as the quantity.

If you set failure_rate to 10. Than 10% of our robots will for sure fail.