Added basic positioning logger
Every n ticks (as specified in the argos file under "log_frequency") every robot will print what x and y that they are currently at.

It is possible for a robot to be at a negative x and negative y value. This is because argos counts x=0 and y=0 relative to the center of the grid.