If light pos is hardcoded then distance can be cal
Currently, in the cpp file I have hardcoded the current position of the light source. Assuming that the hardcoded value matches that in the argos file, then the distance calculation will be correct.

The distance from each robot to the light will be printed when it's x,y position is printed/logged.